Build using catkin when catkin is found
Previously, the package.xml set build_type to "cmake". This meant that
the package could only be built in isolated mode (using
caktin_make_isolated or catkin tools), but not using catkin_make. It
also meant that the ruckig didn't have access to the CATKIN_* cmake
variables that tell it where the compiled ruckig python library should
go in devel and install spaces. After this commit, ruckig tries to
find_package catkin, and if it finds it, then it puts the python library
in the correct place.